% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_operations.R
\name{combineFdObj}
\alias{combineFdObj}
\title{Combine two fdObj objects}
\usage{
combineFdObj(x, y)
}
\arguments{
\item{x}{A `fdObj` object}

\item{y}{A `fdObj` object. This object cannot have any overlapping sample names with the first fdObj.}
}
\value{
A new combined `fdObj` object. metadata matrix will be concatenated using bind_rows (dplyr). rawCurves, approachCurves and retractCurves will be concatenated.
}
\description{
Combine two fdObj objects
}
