% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflection_plot}
\alias{deflection_plot}
\title{Plot deflection vs distance from rawCurves slot of fdObj}
\usage{
deflection_plot(
  obj,
  group_curves_by = NULL,
  split_curves_by = NULL,
  point_size = 0.5,
  alpha = 0.5,
  color_map = NULL
)
}
\arguments{
\item{obj}{A `fdObj` object}

\item{group_curves_by}{Metadata column to color by (optional)}

\item{split_curves_by}{Metadata column to facet by (optional)}

\item{point_size}{Size of points (default = 0.5)}

\item{alpha}{Transparency of points (default = 0.6)}

\item{color_map}{Named vector of colors (optional)}

\item{...}{Additional ggplot2 layers or options}
}
\value{
A ggplot2 object of the scatter plots showing the change of deflection across piezo movement distance. The retraction and extension segments will be plotted in two separate panels.
}
\description{
Plot deflection vs distance from rawCurves slot of fdObj
}
