% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_fd_curves}
\alias{plot_fd_curves}
\title{Plot Force-Distance Curves from an fdObj}
\usage{
plot_fd_curves(
  fdobj,
  curve = "both",
  group_curves_by = NULL,
  split_curves_by = NULL,
  color_map = NULL,
  point_size = 0.5,
  point_alpha = 0.6,
  line_alpha = 0.5,
  text_size = 14,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{fdobj}{An object of class \code{fdObj}.}

\item{curve}{Character. Either "approach", "retract", or "both" (default = "both").}

\item{group_curves_by}{Metadata column to color by (optional).}

\item{split_curves_by}{Metadata column to facet by (optional).}

\item{color_map}{Named vector of colors (optional).}

\item{point_size}{Size of points (default = 0.5).}

\item{point_alpha}{Transparency of points (default = 0.6).}

\item{line_alpha}{Transparency of curve lines (default = 0.5).}

\item{text_size}{Text size for plot labels/ticks/strips/legend (default = 14).}

\item{xlim}{Optional numeric vector of length 2 for x-axis limits.}

\item{ylim}{Optional numeric vector of length 2 for y-axis limits.}

\item{...}{Additional ggplot2 layers/settings to add to the plot
(e.g., \code{theme(...)}, \code{scale_*()}, \code{labs(...)}).}
}
\value{
A ggplot2 object showing the force-distance relationships from AFM curves.
}
\description{
Takes an \code{fdObj} with transformed approach and/or retract curves as
input, extracts the selected transformed curves, and combines them into one
plotting table with separation distance on the x-axis and force on the
y-axis. Optional metadata columns can be used to color curves or split them
across facets, and optional axis limits or extra \code{ggplot2} layers can
be applied. The output is a \code{ggplot2} object showing force-distance
relationships for the requested curves.
}
\examples{
folder <- system.file("extdata", package = "curvana")
fd_obj <- createFdObjFromFolder(folder)
fd_obj <- transform_curves(fd_obj, spring_constant = 0.1, useCurve = "retract", threads = 1, least_length= 300)
plot_fd_curves(fd_obj, curve = "retract")
}
