% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_classes.R
\name{show,fdObj-method}
\alias{show,fdObj-method}
\title{Display a Summary of an fdObj Object}
\usage{
\S4method{show}{fdObj}(object)
}
\arguments{
\item{object}{An \code{fdObj} object.}
}
\value{
Prints a formatted summary of the object in the console.
}
\description{
The `show()` method provides a concise overview of an \code{fdObj} object.
It is automatically called when an fdObj is printed in the console (e.g., typing its name).
The summary includes the number of registered samples, imported raw curves,
successfully transformed approach/retract curves, and metadata column names.
}
\details{
This method is defined for the S4 class \code{fdObj} and is executed automatically
when the object is shown in the console. It provides a quick overview of data
completeness and analysis progress without modifying the object.
}
\examples{
\dontrun{
folder <- system.file("extdata", package = "curvana")
test <- createFdObjFromFolder(folder)
test  # calling 'show()' implicitly
}

}
