% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_analysis.R
\name{analyze_baseline}
\alias{analyze_baseline}
\title{Analyze Baseline from AFM Raw Curves in an fdObj}
\usage{
analyze_baseline(
  fdObj,
  least_length = 150,
  useCurve = NULL,
  slp_threshold = 0.001,
  std_threshold = 0.005,
  threads = 1
)
}
\arguments{
\item{fdObj}{An object of class \code{fdObj}.}

\item{least_length}{Integer. Minimum number of points in the baseline segment.}

\item{useCurve}{Character. Either "approach" or "retract" to specify which raw curve to use.}

\item{slp_threshold}{Numeric. Maximum absolute slope allowed for baseline detection (default = 0.001).}

\item{std_threshold}{Numeric. Maximum standard error of the slope (default = 0.005).}

\item{threads}{Integer. Number of parallel threads to use (default = 1).}
}
\value{
An updated \code{fdObj} with baseline values in the metadata column \code{baseline_nm} and baseline segments in \code{baseline_segment}.
}
\description{
Identifies baseline segments in AFM raw force-distance curves using pre-calculated sensitivity values,
and stores baseline values and segments in the fdObj.
}
