% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_analysis.R
\name{analyze_sensitivity}
\alias{analyze_sensitivity}
\title{Add Sensitivity Estimates to an fdObj}
\usage{
analyze_sensitivity(
  fdObj,
  end = 200,
  intv = 4,
  R_squared_min = 0.99,
  minimum_length = 4,
  useCurve = "approach",
  threads = 1
)
}
\arguments{
\item{fdObj}{An object of class \code{fdObj}.}

\item{end}{Integer. The maximum index in raw curves to consider (e.g., 200).}

\item{intv}{Integer. Chunk size for sensitivity calculation (e.g., 4).}

\item{R_squared_min}{Numeric. Minimum \eqn{R^2} threshold passed to
\code{calc_sensitivity()}. Default is 0.99.}

\item{minimum_length}{Integer. Minimum accumulated points required for a valid sensitivity
result. Passed to \code{calc_sensitivity()}. Default is 4.}

\item{useCurve}{Character. Either "approach" or "retract" to determine which curve to use.}

\item{threads}{Number of parallel threads to use (default = 1).}
}
\value{
An updated \code{fdObj} with sensitivity values in metadata and segments in senscal_segment.
}
\description{
Applies sensitivity estimation to each rawCurve in an fdObj and stores the sensitivity values
in the metadata and the segment used in the senscal_segment slot.
}
