% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_analysis.R
\name{calc_sensitivity}
\alias{calc_sensitivity}
\title{Calculate Sensitivity from a Segment of AFM Data}
\usage{
calc_sensitivity(end, intv, x, y)
}
\arguments{
\item{end}{Integer. The maximum index to consider in the `x` and `y` vectors (i.e., up to which point to search).}

\item{intv}{Integer. The chunk size to use when iteratively selecting data segments.}

\item{x}{Numeric vector. The piezo extension values (e.g., distance in nm).}

\item{y}{Numeric vector. The deflection values (e.g., voltage).}
}
\value{
A list of three elements:
\describe{
  \item{senscal_segment}{a dataframe of a subset of points in the original raw curve which are used for sensitivity calibration}
  \item{sensitivity}{The calculated sensitivity (slope) as a numeric value, or \code{NULL} if not computable.}
}
}
\description{
This function identifies a linear segment from AFM deflection vs. piezo extension data and calculates
the sensitivity (slope) using linear regression. It iteratively adds chunks of data to build a segment
with high linearity (\eqn{R^2 > 0.999}) and optionally removes outliers to maintain this criterion.
}
\examples{
x <- seq(0, 100, by = 1)
y <- 0.02 * x + rnorm(length(x), sd = 0.01)
calc_sensitivity(end = 80, intv = 10, x = x, y = y)

}
