% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_operations.R
\name{combineFdObj}
\alias{combineFdObj}
\title{Combine two fdObj objects}
\usage{
combineFdObj(x, y)
}
\arguments{
\item{x}{A `fdObj` object}

\item{y}{A `fdObj` object. This object cannot have any overlapping sample names with the first fdObj.}
}
\value{
A new combined `fdObj` object. metadata matrix will be concatenated using bind_rows (dplyr). All slots will be concatenated.
}
\description{
Combine two fdObj objects
}
\examples{
folder <- system.file("extdata", package = "curvana")
fd_obj <- createFdObjFromFolder(folder)
md <- fd_obj@metadata
surface <- sub("_.*", "", md$filename)
surface[surface == "F"] <- "Group_A"
surface[surface == "P"] <- "Group_B"
surface[surface == "silicon"] <- "Group_C"
surface[surface == "Z"] <- "Group_D"
md$surface <- surface
fd_obj@metadata <- md
fd_obj2 = extract(fd_obj,by_col = list(surface = 'Group_B'))
fd_obj3 = extract(fd_obj,by_sample = 'F_1.spm-F4E1_ForceCurveIndex_0.spm')
fd_obj4 = combineFdObj(fd_obj2, fd_obj3)
}
