% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_classes.R
\docType{class}
\name{fdObj-class}
\alias{fdObj-class}
\title{An S4 class to hold AFM force-distance raw and processed curves with metadata}
\description{
An S4 class to hold AFM force-distance raw and processed curves with metadata
}
\section{Slots}{

\describe{
\item{\code{rawCurves}}{A named list of data.frames containing raw AFM data. Each
data.frame must include columns \code{Calc_Ramp_Ex_nm}, \code{Calc_Ramp_Rt_nm},
\code{Defl_V_Ex}, and \code{Defl_V_Rt}; additional columns are allowed.}

\item{\code{approachCurves}}{A named list of 2-column data.frames (distance vs force from approach segment)}

\item{\code{retractCurves}}{A named list of 2-column data.frames (distance vs force from retract segment)}

\item{\code{metadata}}{A data.frame with row names matching names(rawCurves). Stores metadata and calculated features.}

\item{\code{senscal_segment}}{A list with two slots: approach and retract, each is a named list of data.frames, storing the part of the raw curve used for sensitivity calculation.}

\item{\code{baseline_segment}}{A list with two slots: approach and retract, each is a named list of data.frames, storing the part of the raw curve used for baseline calculation}
}}

