% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_deflection_curves}
\alias{plot_deflection_curves}
\title{Plot raw deflection vs piezo distance from rawCurves slot of fdObj}
\usage{
plot_deflection_curves(
  fdobj,
  curve = c("both", "approach", "retract"),
  group_curves_by = NULL,
  split_curves_by = NULL,
  point_size = 0.5,
  alpha = 0.5,
  line_alpha = 0.5,
  color_map = NULL
)
}
\arguments{
\item{fdobj}{A `fdObj` object}

\item{curve}{Which segment(s) to plot: "approach", "retract", or "both"}

\item{group_curves_by}{Metadata column to color by (optional)}

\item{split_curves_by}{Metadata column to facet by (optional)}

\item{point_size}{Size of points (default = 0.5)}

\item{alpha}{Transparency of points (default = 0.6)}

\item{line_alpha}{Transparency of connecting paths (default = 0.5)}

\item{color_map}{Named vector of colors (optional)}
}
\value{
A ggplot2 object of the scatter plots showing raw deflection signal.
}
\description{
Plot raw deflection vs piezo distance from rawCurves slot of fdObj
}
