% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_fd_curves}
\alias{plot_fd_curves}
\title{Plot Force-Distance Curves from an fdObj}
\usage{
plot_fd_curves(
  fdobj,
  curve = "both",
  group_curves_by = NULL,
  split_curves_by = NULL,
  color_map = NULL,
  point_size = 0.5,
  point_alpha = 0.6,
  line_alpha = 0.5,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{fdobj}{An object of class \code{fdObj}.}

\item{curve}{Character. Either "approach", "retract", or "both" (default = "both").}

\item{group_curves_by}{Metadata column to color by (optional).}

\item{split_curves_by}{Metadata column to facet by (optional).}

\item{color_map}{Named vector of colors (optional).}

\item{point_size}{Size of points (default = 0.5).}

\item{point_alpha}{Transparency of points (default = 0.6).}

\item{line_alpha}{Transparency of curve lines (default = 0.5).}

\item{xlim}{Optional numeric vector of length 2 for x-axis limits.}

\item{ylim}{Optional numeric vector of length 2 for y-axis limits.}

\item{...}{Additional ggplot2 layers/settings to add to the plot
(e.g., \code{theme(...)}, \code{scale_*()}, \code{labs(...)}).}
}
\value{
A ggplot2 object showing the force-distance relationships from AFM curves.
}
\description{
Generates force-distance plots from the calibrated approach or retract curves in an \code{fdObj}.
}
