% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_analysis.R
\name{transform_curves}
\alias{transform_curves}
\title{Transform All Curves in an fdObj into Separation Distance and Force}
\usage{
transform_curves(fdObj, spring_constant, useCurve = "approach", threads = 1)
}
\arguments{
\item{fdObj}{An object of class \code{fdObj}.}

\item{spring_constant}{Numeric. Spring constant in nN/nm.}

\item{useCurve}{Character. Either "approach" or "retract".}

\item{threads}{Integer. Number of parallel threads to use (default = 1).}
}
\value{
An updated \code{fdObj} with transformed curves stored in the corresponding slot.
}
\description{
Applies `transform_a_curve()` to all raw curves in an fdObj object.
}
